requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
APScheduler==3.10.4
email-validator==2.1.0
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0